collisions fix

+++ include/vector.h
@@ -2,13 +2,13 @@
 #define _VECTOR_H
 
 struct vector {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 struct vector vector_add(struct vector first, struct vector second);
 
-struct vector vector_multiply(struct vector vec, int scalar);
+struct vector vector_multiply(struct vector vec, float scalar);
 
 #endif
