+++ src/lib/vector.c
@@ -6,7 +6,7 @@ struct vector vector_add(struct vector first, struct vector second) {
     return first;
 }
 
-struct vector vector_multiply(struct vector vec, int scalar) {
+struct vector vector_multiply(struct vector vec, float scalar) {
     vec.x *= scalar;
     vec.y *= scalar;
     return vec;
